requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
Office365-REST-Python-Client==2.5.3
